Got clipboard support working! All the changes are in luigi.h, and I've put "//@clipboard_change" near all the changes to make it easy to find them.
There is a little catch though, X11 makes it so that when you paste something it creates an event that the program you are pasting from has to respond to by sending over the data. Which means that in order to paste something you copied in gf you have to keep that text in memory. I'm not sure what your general policy for keeping track of that kind of stuff is, so for now I just created a giant global char array, and you can adapt that however you want.

Also when you close gf you lose whatever was in the clipboard (since the program is no longer there to copy from). A lot of linux programs just work this way, and the ones that don't just keep a process running in the background (or try to find a clipboard manager that continuously runs in the background) until the clipboard is replaced. I just went with the "close the program -> lose the text" approach.