feat: added the ability to enable, disable and delete selected breakpoints.
This commit adds the ability to delete, enable and disable multiple
selected breakpoints at once.
The user can select multiple breakpoints by holding the CTRL key and
clicking with the left mouse button on the desired breakpoints in the
`Breakpoint` window. Then if they choose in the right click menu to
`Delete`, `Enable` or `Disable`, the action will affect all the selected
breakpoints.