fix: handle conditional breakpoints correctly.
This commit adds the ability to have multiple breakpoints at the same
line if they have different conditions. To distinguish between them we
have added a new column in the `Breakpoint` window to display the
condition for each breakpoint.

Deleting, Disabling or Enabling a breakpoint from the marker on the left
margin of the code view, will Delete, Disable or Enable all the
breakpoints at that line accordingly.

The breakpoint marker in the left margin of the code view will only show
in a disabled state if all the breakpoints points at that given line are
disabled.