Also set gdbArgv[0] when loading gdb path from settings
This is necessary for rust-gdb to work since that is just a symlink to
rustup which will select the correct rust-gdb installation at runtime.
(Technically, you try to find the path to the actual rust-gdb binary
somewhere in ~/.rustup and use that).

In addition, setting argv[0] to the program filename is what most
programs should expect anyway.